refactor: #81 - LegalContent 컴포넌트 분리
